cpp: Start adding Complex, Set, and Double

+++ cpp/include/mamba/builtins/__concepts/hashable.hpp
@@ -3,13 +3,13 @@
 #include <concepts>
 
 #include "mamba/builtins/__concepts/value.hpp"
-#include "mamba/builtins/__types/int.hpp"
+#include "mamba/builtins/__types/big_int.hpp"
 
 namespace mamba::builtins::details {
 
 template <typename T>
 concept HashableObject = requires(const T t) {
-  { t.__Hash__() } -> std::same_as<Int>;
+  { t.__Hash__() } -> std::same_as<BigInt>;
 };
 
 template <typename T>
